Start adding support for masked values to parafac()
The idea to modify the tensor with the masked array is from Tomasi, Giorgio,
and Rasmus Bro. "PARAFAC and missing values." Chemometrics and Intelligent
Laboratory Systems 75.2 (2005): 163-180.

This still needs to be modified to continue to work with sparse arrays. Right
now, it fully decomposes the factors to compute the modification, which will
not work if the fully decomposed factors do not fit in memory.