Make the parafac mask support sparse friendly
This required adding a mask flag to kruskal_to_tensor() and kr() so that the
mask could be absorbed into the calculation in a sparse friendly way. In order
to be sparse friendly, the mask should be a sparse array with a fill value of
1 (True). In other words, the number of missing values should itself be
sparse.

The generic kr() algorithm was rewritten to use a generic outer product, and
kruskal_to_tensor() now uses sum instead of dot (the NumPy kr() just
multiplies the mask).