Fix the tucker tests for sparse
The tests now test both init='random' and init='svd'.

The tucker algorithms unfortunately fail in the sparse backend if the ranks
are too large, including the default rank=None which chooses the shape of the
input tensor. This is because scipy.sparse.linalg.eigsh fails if asked for too
many eigenvalues. I do not know if this can be fixed. For now, I am skipping
the rank=None tests in the sparse backend.

I have reduced the ranks in some of the tests here so that they can be run
with the sparse algorithm.