Fix torch tensor creation dtype/device
The previous implementation simply took the input data and forced it to
be on the given device (`cpu` default) with the given dtype (`float32`
default).

This lead to `tl.tensor(torch.zeros(1, device="cuda:0"))` creating a
float32 tensor on `cpu` while the expected behavior should be an
int32 tensor on `cuda:0`.

Resolves: #537